Move construction of domain participant entity to level above
