+++ dds/src/implementation/domain_participant_backend/entities/domain_participant.rs
@@ -61,6 +61,7 @@ use super::{publisher::PublisherEntity, subscriber::SubscriberEntity, topic::Top
 
 pub struct DomainParticipantEntity {
     domain_id: DomainId,
+    instance_handle: InstanceHandle,
     qos: DomainParticipantQos,
     builtin_subscriber: SubscriberEntity,
     builtin_publisher: PublisherEntity,
@@ -89,256 +90,17 @@ impl DomainParticipantEntity {
         domain_participant_qos: DomainParticipantQos,
         listener: Option<Box<dyn DomainParticipantListenerAsync + Send>>,
         status_kind: Vec<StatusKind>,
-        instance_handle_counter: &mut InstanceHandleCounter,
-        executor: &Executor,
-        transport: &mut dyn Transport,
+        status_condition: Actor<StatusConditionActor>,
+        instance_handle: InstanceHandle,
+        builtin_publisher: PublisherEntity,
+        builtin_subscriber: SubscriberEntity,
+        topic_list: HashMap<String, TopicEntity>,
     ) -> Self {
-        fn sedp_data_reader_qos() -> DataReaderQos {
-            DataReaderQos {
-                durability: DurabilityQosPolicy {
-                    kind: DurabilityQosPolicyKind::TransientLocal,
-                },
-                history: HistoryQosPolicy {
-                    kind: HistoryQosPolicyKind::KeepLast(1),
-                },
-                reliability: ReliabilityQosPolicy {
-                    kind: ReliabilityQosPolicyKind::Reliable,
-                    max_blocking_time: DurationKind::Finite(Duration::new(0, 0)),
-                },
-                ..Default::default()
-            }
-        }
-
-        fn sedp_data_writer_qos() -> DataWriterQos {
-            DataWriterQos {
-                durability: DurabilityQosPolicy {
-                    kind: DurabilityQosPolicyKind::TransientLocal,
-                },
-                history: HistoryQosPolicy {
-                    kind: HistoryQosPolicyKind::KeepLast(1),
-                },
-                reliability: ReliabilityQosPolicy {
-                    kind: ReliabilityQosPolicyKind::Reliable,
-                    max_blocking_time: DurationKind::Finite(Duration::new(0, 0)),
-                },
-                ..Default::default()
-            }
-        }
-
-        let mut topic_list = HashMap::new();
-        let spdp_topic_participant_handle = instance_handle_counter.generate_new_instance_handle();
-
-        let mut spdp_topic_participant = TopicEntity::new(
-            TopicQos::default(),
-            "SpdpDiscoveredParticipantData".to_string(),
-            DCPS_PARTICIPANT.to_owned(),
-            spdp_topic_participant_handle,
-            Actor::spawn(StatusConditionActor::default(), &executor.handle()),
-            None,
-            vec![],
-            Arc::new(SpdpDiscoveredParticipantData::get_type()),
-        );
-        spdp_topic_participant.enable();
-
-        topic_list.insert(DCPS_PARTICIPANT.to_owned(), spdp_topic_participant);
-
-        let sedp_topic_topics_handle = instance_handle_counter.generate_new_instance_handle();
-        let mut sedp_topic_topics = TopicEntity::new(
-            TopicQos::default(),
-            "DiscoveredTopicData".to_string(),
-            DCPS_TOPIC.to_owned(),
-            sedp_topic_topics_handle,
-            Actor::spawn(StatusConditionActor::default(), &executor.handle()),
-            None,
-            vec![],
-            Arc::new(DiscoveredTopicData::get_type()),
-        );
-        sedp_topic_topics.enable();
-
-        topic_list.insert(DCPS_TOPIC.to_owned(), sedp_topic_topics);
-
-        let sedp_topic_publications_handle = instance_handle_counter.generate_new_instance_handle();
-        let mut sedp_topic_publications = TopicEntity::new(
-            TopicQos::default(),
-            "DiscoveredWriterData".to_string(),
-            DCPS_PUBLICATION.to_owned(),
-            sedp_topic_publications_handle,
-            Actor::spawn(StatusConditionActor::default(), &executor.handle()),
-            None,
-            vec![],
-            Arc::new(DiscoveredWriterData::get_type()),
-        );
-        sedp_topic_publications.enable();
-        topic_list.insert(DCPS_PUBLICATION.to_owned(), sedp_topic_publications);
-
-        let sedp_topic_subscriptions_handle =
-            instance_handle_counter.generate_new_instance_handle();
-        let mut sedp_topic_subscriptions = TopicEntity::new(
-            TopicQos::default(),
-            "DiscoveredReaderData".to_string(),
-            DCPS_SUBSCRIPTION.to_owned(),
-            sedp_topic_subscriptions_handle,
-            Actor::spawn(StatusConditionActor::default(), &executor.handle()),
-            None,
-            vec![],
-            Arc::new(DiscoveredReaderData::get_type()),
-        );
-        sedp_topic_subscriptions.enable();
-        topic_list.insert(DCPS_SUBSCRIPTION.to_owned(), sedp_topic_subscriptions);
-
-        let spdp_writer_qos = DataWriterQos {
-            durability: DurabilityQosPolicy {
-                kind: DurabilityQosPolicyKind::TransientLocal,
-            },
-            history: HistoryQosPolicy {
-                kind: HistoryQosPolicyKind::KeepLast(1),
-            },
-            reliability: ReliabilityQosPolicy {
-                kind: ReliabilityQosPolicyKind::BestEffort,
-                max_blocking_time: DurationKind::Finite(Duration::new(0, 0)),
-            },
-            ..Default::default()
-        };
-        let spdp_reader_qos = DataReaderQos {
-            durability: DurabilityQosPolicy {
-                kind: DurabilityQosPolicyKind::TransientLocal,
-            },
-            history: HistoryQosPolicy {
-                kind: HistoryQosPolicyKind::KeepLast(1),
-            },
-            reliability: ReliabilityQosPolicy {
-                kind: ReliabilityQosPolicyKind::BestEffort,
-                max_blocking_time: DurationKind::Finite(Duration::new(0, 0)),
-            },
-            ..Default::default()
-        };
-
-        let dcps_participant_reader = DataReaderEntity::new(
-            instance_handle_counter.generate_new_instance_handle(),
-            spdp_reader_qos,
-            topic_list[DCPS_PARTICIPANT].topic_name().to_owned(),
-            topic_list[DCPS_PARTICIPANT].type_name().to_owned(),
-            topic_list[DCPS_PARTICIPANT].type_support().clone(),
-            Actor::spawn(StatusConditionActor::default(), &executor.handle()),
-            None,
-            Vec::new(),
-            transport.get_participant_discovery_reader(),
-        );
-        let dcps_topic_reader = DataReaderEntity::new(
-            instance_handle_counter.generate_new_instance_handle(),
-            sedp_data_reader_qos(),
-            topic_list[DCPS_TOPIC].topic_name().to_owned(),
-            topic_list[DCPS_TOPIC].type_name().to_owned(),
-            topic_list[DCPS_TOPIC].type_support().clone(),
-            Actor::spawn(StatusConditionActor::default(), &executor.handle()),
-            None,
-            Vec::new(),
-            transport.get_topics_discovery_reader(),
-        );
-        let dcps_publication_reader = DataReaderEntity::new(
-            instance_handle_counter.generate_new_instance_handle(),
-            sedp_data_reader_qos(),
-            topic_list[DCPS_PUBLICATION].topic_name().to_owned(),
-            topic_list[DCPS_PUBLICATION].type_name().to_owned(),
-            topic_list[DCPS_PUBLICATION].type_support().clone(),
-            Actor::spawn(StatusConditionActor::default(), &executor.handle()),
-            None,
-            Vec::new(),
-            transport.get_topics_discovery_reader(),
-        );
-        let dcps_subscription_reader = DataReaderEntity::new(
-            instance_handle_counter.generate_new_instance_handle(),
-            sedp_data_reader_qos(),
-            topic_list[DCPS_SUBSCRIPTION].topic_name().to_owned(),
-            topic_list[DCPS_SUBSCRIPTION].type_name().to_owned(),
-            topic_list[DCPS_SUBSCRIPTION].type_support().clone(),
-            Actor::spawn(StatusConditionActor::default(), &executor.handle()),
-            None,
-            Vec::new(),
-            transport.get_topics_discovery_reader(),
-        );
-
-        let mut builtin_subscriber = SubscriberEntity::new(
-            instance_handle_counter.generate_new_instance_handle(),
-            SubscriberQos::default(),
-            Actor::spawn(StatusConditionActor::default(), &executor.handle()),
-            None,
-            vec![],
-        );
-        builtin_subscriber.enable();
-        builtin_subscriber.insert_data_reader(dcps_participant_reader);
-        builtin_subscriber.insert_data_reader(dcps_topic_reader);
-        builtin_subscriber.insert_data_reader(dcps_publication_reader);
-        builtin_subscriber.insert_data_reader(dcps_subscription_reader);
-
-        let mut dcps_participant_writer = DataWriterEntity::new(
-            instance_handle_counter.generate_new_instance_handle(),
-            transport.get_participant_discovery_writer(),
-            topic_list[DCPS_PARTICIPANT].topic_name().to_owned(),
-            topic_list[DCPS_PARTICIPANT].type_name().to_owned(),
-            topic_list[DCPS_PARTICIPANT].type_support().clone(),
-            Actor::spawn(StatusConditionActor::default(), &executor.handle()),
-            None,
-            vec![],
-            spdp_writer_qos,
-        );
-        dcps_participant_writer.enable();
-
-        let mut dcps_topics_writer = DataWriterEntity::new(
-            instance_handle_counter.generate_new_instance_handle(),
-            transport.get_topics_discovery_writer(),
-            topic_list[DCPS_TOPIC].topic_name().to_owned(),
-            topic_list[DCPS_TOPIC].type_name().to_owned(),
-            topic_list[DCPS_TOPIC].type_support().clone(),
-            Actor::spawn(StatusConditionActor::default(), &executor.handle()),
-            None,
-            vec![],
-            sedp_data_writer_qos(),
-        );
-        dcps_topics_writer.enable();
-        let mut dcps_publications_writer = DataWriterEntity::new(
-            instance_handle_counter.generate_new_instance_handle(),
-            transport.get_publications_discovery_writer(),
-            topic_list[DCPS_PUBLICATION].topic_name().to_owned(),
-            topic_list[DCPS_PUBLICATION].type_name().to_owned(),
-            topic_list[DCPS_PUBLICATION].type_support().clone(),
-            Actor::spawn(StatusConditionActor::default(), &executor.handle()),
-            None,
-            vec![],
-            sedp_data_writer_qos(),
-        );
-        dcps_publications_writer.enable();
-
-        let mut dcps_subscriptions_writer = DataWriterEntity::new(
-            instance_handle_counter.generate_new_instance_handle(),
-            transport.get_subscriptions_discovery_writer(),
-            topic_list[DCPS_SUBSCRIPTION].topic_name().to_owned(),
-            topic_list[DCPS_SUBSCRIPTION].type_name().to_owned(),
-            topic_list[DCPS_SUBSCRIPTION].type_support().clone(),
-            Actor::spawn(StatusConditionActor::default(), &executor.handle()),
-            None,
-            vec![],
-            sedp_data_writer_qos(),
-        );
-        dcps_subscriptions_writer.enable();
-        let mut builtin_publisher = PublisherEntity::new(
-            PublisherQos::default(),
-            instance_handle_counter.generate_new_instance_handle(),
-            None,
-            vec![],
-            Actor::spawn(StatusConditionActor::default(), &executor.handle()),
-        );
-        builtin_publisher.enable();
-        builtin_publisher.insert_data_writer(dcps_participant_writer);
-        builtin_publisher.insert_data_writer(dcps_topics_writer);
-        builtin_publisher.insert_data_writer(dcps_publications_writer);
-        builtin_publisher.insert_data_writer(dcps_subscriptions_writer);
-
         let participant_listener_thread = listener.map(ParticipantListenerThread::new);
-        let status_condition = Actor::spawn(StatusConditionActor::default(), &executor.handle());
 
         Self {
             domain_id,
+            instance_handle,
             qos: domain_participant_qos,
             builtin_subscriber,
             builtin_publisher,
@@ -369,9 +131,8 @@ impl DomainParticipantEntity {
         self.enabled = true;
     }
 
-    pub fn get_instance_handle(&self) -> InstanceHandle {
-        todo!()
-        // InstanceHandle::new(self.transport.guid().into())
+    pub fn instance_handle(&self) -> InstanceHandle {
+        self.instance_handle
     }
 
     pub fn get_statuscondition(&self) -> ActorAddress<StatusConditionActor> {
